+++ Dockerfile
@@ -1,16 +1,21 @@
-FROM rust:1.45-slim-stretch
+FROM rust:1.47-slim-buster
 
 ARG DEBIAN_FRONTEND=noninterative
 
-ARG SCCACHE_URL=https://github.com/mozilla/sccache/releases/download/0.2.12/sccache-0.2.12-x86_64-unknown-linux-musl.tar.gz
+ARG SCCACHE_BASE=https://github.com/mozilla/sccache/releases/download
+ARG SCCACHE_VER=0.2.13
+ARG SCCACHE_URL=${SCCACHE_BASE}/${SCCACHE_VER}/sccache-${SCCACHE_VER}-x86_64-unknown-linux-musl.tar.gz
 
 ENV SCCACHE_CACHE_SIZE=100G
 ENV SCCACHE_DIR=/opt/sccache
 
 RUN mkdir -p $SCCACHE_DIR
 
-RUN    echo "deb http://deb.debian.org/debian stretch-backports main non-free" >/etc/apt/sources.list.d/backports.list \
-    && echo "deb-src http://httpredir.debian.org/debian stretch-backports main non-free" >>/etc/apt/sources.list.d/backports.list
+ARG CARGO_MAKE_VER=0.32.7
+ARG CARGO_MAKE_DIR=cargo-make-v${CARGO_MAKE_VER}-x86_64-unknown-linux-musl
+ARG CARGO_MAKE_FILE=${CARGO_MAKE_DIR}.zip
+ARG CARGO_MAKE_URL_BASE=https://github.com/sagiegurari/cargo-make/releases/download/
+ARG CARGO_MAKE_URL=${CARGO_MAKE_URL_BASE}/${CARGO_MAKE_VER}/${CARGO_MAKE_FILE}
 
 RUN \
       apt-get update \
@@ -24,7 +29,12 @@ RUN \
    && rm -rf /tmp/sccache /tmp/sccache.tgz \
    && rustup target add x86_64-unknown-linux-musl \
    && rustup component add rustfmt \
-   && rustup component add clippy
+   && rustup component add clippy \
+   && curl -sSLO ${CARGO_MAKE_URL} \
+   && unzip ${CARGO_MAKE_FILE} \
+   && mv ${CARGO_MAKE_DIR}/cargo-make /usr/local/bin \
+   && chmod +x /usr/local/bin/cargo-make \
+   && rm -rf ${CARGO_MAKE_DIR} ${CARGO_MAKE_FILE}
 
 ENV RUSTC_WRAPPER=/usr/local/bin/sccache
